updated admin code so there's a cmd admin folder and separate admin lib
